refactor: componentize button and dropdown
